local-agent/pyproject.toml
[project]
name = "local-agent"
version = "0.2.1"
description = "Local RAG agent with FastAPI, LangGraph, SQLite, Chroma"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "aiosqlite>=0.20",
    "uvicorn>=0.30",
    "pydantic>=2.7",
    "sqlalchemy>=2.0",
    "chromadb>=0.5",
    "tiktoken>=0.7",
    "langgraph>=0.2",
    "langgraph-cli[inmem]>=0.1.55",
    "langchain-core>=0.2",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "openai>=1.35",
    "requests>=2.32.4",
    "pypdf>=4.2",
    "python-docx>=1.1",
    "python-pptx>=0.6.23",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
]

[tool.uv]
package = true

[tool.pyright]
typeCheckingMode = "basic"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
